Utils - Hydratator - Fix WP_User field ACF
When using User field inside ACF, the method get_field return an array representing the user rather than a WP_User.
This behaviour prevent the returned value to be Hydrated by Hydratator.
Hydratator now detect early if we try to hydrate an array representing a user by verify if user_registered and user_avatar key are present in array. If this the case, we transform this array into a WP_User.

* Fix issue with User ACF Field returned value and the hydrate process